Rounding vertices in RoadNetwork
The reason for this is that if many decimal places are uses, it is very unlikely for vertices to be equal and cycles will pretty much never be found. Now the level of detail can be specified by setting the number of decimal places for each vertex when instantiating RoadNetwork. The number decimal places has to be in the range [0,15] where 15 gives a lot of detail. I set the default to 1 since it looked quite good and makes it more likely for cycles to be found in the road network.