Bug fixes to make concurrency work
UnityEngine.Random.Range seemingly can't be called on other threads than the main thread. Inclusive float ranges are tricky to do yourself, so I just dispatched the work to the main thread. If we figure out how to do it ourselves, that would be a nicer solution.

UnityEngine.Random.Range(0, 1) always returns 0 (since it is half closed and returns an integer). I changed it to new Random().Next(0, 2) which will either retun 0 or 1.